# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RenderHealthStatus = typeof RenderHealthStatus[keyof typeof RenderHealthStatus];


export const RenderHealthStatus = {
  ok: 'ok',
  degraded: 'degraded',
} as const;

export interface RenderHealth {
  status: RenderHealthStatus;
  ffmpeg: boolean;
  supabase: boolean;
  worker: boolean;
  queueDepth: number;
}

export type ImageInputContentType = typeof ImageInputContentType[keyof typeof ImageInputContentType];


export const ImageInputContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface ImageInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  contentType: ImageInputContentType;
  /** @minLength 32 */
  dataUrl: string;
}

export interface SourceInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @minItems 1
     * @maxItems 100
     */
  images: ImageInput[];
}

export interface SourceImage {
  id: string;
  name: string;
  contentType: string;
  previewUrl: string;
  width: number;
  height: number;
}

export interface Source {
  id: string;
  name: string;
  imageCount: number;
  images: SourceImage[];
  createdAt: string;
}

export type VideoFormat = typeof VideoFormat[keyof typeof VideoFormat];


export const VideoFormat = {
  '9:16': '9:16',
  '1:1': '1:1',
  '16:9': '16:9',
} as const;

export type Quality = typeof Quality[keyof typeof Quality];


export const Quality = {
  draft: 'draft',
  high: 'high',
  maximum: 'maximum',
} as const;

export type SequenceMode = typeof SequenceMode[keyof typeof SequenceMode];


export const SequenceMode = {
  random: 'random',
  sequential: 'sequential',
} as const;

export type RenderInputFps = typeof RenderInputFps[keyof typeof RenderInputFps];


export const RenderInputFps = {
  NUMBER_30: 30,
  NUMBER_60: 60,
} as const;

export interface RenderInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  sourceId: string;
  /**
     * @minimum 0.5
     * @maximum 1200
     */
  duration: number;
  /**
     * @minimum 0.03
     * @maximum 10
     */
  photoDuration: number;
  format: VideoFormat;
  fps: RenderInputFps;
  quality: Quality;
  sequence: SequenceMode;
  /**
     * @minimum 0
     * @maximum 2147483647
     * @nullable
     */
  seed?: number | null;
  /** @nullable */
  callbackUrl?: string | null;
  /**
     * @maxLength 120
     * @nullable
     */
  filename?: string | null;
}

export type RenderAcceptedStatus = typeof RenderAcceptedStatus[keyof typeof RenderAcceptedStatus];


export const RenderAcceptedStatus = {
  queued: 'queued',
  rendering: 'rendering',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface RenderAccepted {
  jobId: string;
  status: RenderAcceptedStatus;
  seed: number;
}

export interface RenderOutput {
  url: string;
  filename: string;
  mimeType: string;
  size: number;
  width: number;
  height: number;
  fps: number;
  duration: number;
  frameCount: number;
  codec: string;
  hasAudio: boolean;
}

export type RenderJobStatus = typeof RenderJobStatus[keyof typeof RenderJobStatus];


export const RenderJobStatus = {
  queued: 'queued',
  rendering: 'rendering',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface RenderError {
  code: string;
  message: string;
}

export interface RenderJob {
  jobId: string;
  status: RenderJobStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  stage: string;
  seed: number;
  elapsedSeconds: number;
  /** @nullable */
  estimatedRemainingSeconds: number | null;
  output: RenderOutput | null;
  error: RenderError | null;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export type IdempotencyKeyParameter = string;