# lib/api-zod/src/generated/types/imageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageInputContentType } from './imageInputContentType';

export interface ImageInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  contentType: ImageInputContentType;
  /** @minLength 32 */
  dataUrl: string;
}