# lib/api-zod/src/generated/types/quality.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Quality = typeof Quality[keyof typeof Quality];


export const Quality = {
  draft: 'draft',
  high: 'high',
  maximum: 'maximum',
} as const;