# lib/api-zod/src/generated/types/renderAccepted.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RenderAcceptedStatus } from './renderAcceptedStatus';

export interface RenderAccepted {
  jobId: string;
  status: RenderAcceptedStatus;
  seed: number;
}