# lib/api-zod/src/generated/types/renderHealth.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RenderHealthStatus } from './renderHealthStatus';

export interface RenderHealth {
  status: RenderHealthStatus;
  ffmpeg: boolean;
  supabase: boolean;
  worker: boolean;
  queueDepth: number;
}