# lib/api-zod/src/generated/types/renderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Quality } from './quality';
import type { RenderInputFps } from './renderInputFps';
import type { SequenceMode } from './sequenceMode';
import type { VideoFormat } from './videoFormat';

export interface RenderInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  sourceId: string;
  /**
     * @minimum 0.5
     * @maximum 1200
     */
  duration: number;
  /**
     * @minimum 0.03
     * @maximum 10
     */
  photoDuration: number;
  format: VideoFormat;
  fps: RenderInputFps;
  quality: Quality;
  sequence: SequenceMode;
  /**
     * @minimum 0
     * @maximum 2147483647
     * @nullable
     */
  seed?: number | null;
  /** @nullable */
  callbackUrl?: string | null;
  /**
     * @maxLength 120
     * @nullable
     */
  filename?: string | null;
}