# lib/api-zod/src/generated/types/renderJob.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RenderError } from './renderError';
import type { RenderJobStatus } from './renderJobStatus';
import type { RenderOutput } from './renderOutput';

export interface RenderJob {
  jobId: string;
  status: RenderJobStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  stage: string;
  seed: number;
  elapsedSeconds: number;
  /** @nullable */
  estimatedRemainingSeconds: number | null;
  output: RenderOutput | null;
  error: RenderError | null;
}