# lib/api-zod/src/generated/types/source.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceImage } from './sourceImage';

export interface Source {
  id: string;
  name: string;
  imageCount: number;
  images: SourceImage[];
  createdAt: Date;
}