# lib/api-zod/src/generated/types/sourceImage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SourceImage {
  id: string;
  name: string;
  contentType: string;
  previewUrl: string;
  width: number;
  height: number;
}