# lib/api-zod/src/generated/types/sourceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageInput } from './imageInput';

export interface SourceInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @minItems 1
     * @maxItems 100
     */
  images: ImageInput[];
}