# lib/api-zod/src/generated/types/videoFormat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type VideoFormat = typeof VideoFormat[keyof typeof VideoFormat];


export const VideoFormat = {
  '9:16': '9:16',
  '1:1': '1:1',
  '16:9': '16:9',
} as const;